vertex_ai/adk_kor/requirements.txt
# Core dependencies
google-adk>=0.1.0
python-dotenv>=1.0.0
requests>=2.31.0

# Optional dependencies for enhanced functionality
Pillow>=10.0.0  # For image processing
